+++ lib/src/wheel_picker_controller.dart
@@ -1,3 +1,5 @@
+import 'dart:math' as math;
+
 import 'package:flutter/material.dart';
 import './wheel_picker_style.dart';